pyproject.toml
[project]
name = "mcp-jira"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.27.0",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
